updateVueServerRender
Vue-server-render 2.6.10 uses serialize-javascript version 1.5.0, which comes with a vulnerability already known and published in CVE:
CVE-2019-16769	The serialize-javascript npm package before version 2.1.1 is vulnerable to Cross-site Scripting (XSS). It does not properly mitigate against unsafe characters in serialized regular expressions. This vulnerability is not affected on Node.js environment since Node.js's implementation of RegExp.prototype.toString() backslash-escapes all forward slashes in regular expressions. If serialized data of regular expression objects are used in an environment other than Node.js, it is affected by this vulnerability.

The problem has been solved in later versions, so updating vue-server-render from 2.6.10 to 2.6.11 will automatically fix this serious issue, as it uses a newer version of serialize-javascript.